Rename build -> compile
